Changes as per Task 2

+++ ImageHoster/src/main/java/ImageHoster/controller/ImageController.java
@@ -91,14 +91,19 @@ public class ImageController {
 
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
-    @RequestMapping(value = "images/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    @RequestMapping(value = "/editImage")
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
         Image image = imageService.getImage(imageId);
-
-        //String tags = convertTagsToString(image.getTags());
+        User user = (User) session.getAttribute("loggeduser");
+        String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
-        //model.addAttribute("tags", tags);
-        return "/images/{title}";
+        model.addAttribute("tags", tags);
+        if(image.getUser().getId()!=user.getId()){
+            model.addAttribute("editError",true);
+            return "images/image";}
+        else{
+            return "images/edit";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,7 +145,15 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model,HttpSession session) {
+        User user = (User) session.getAttribute("loggeduser");
+        Image image = imageService.getImage(imageId);
+        String tags = convertTagsToString(image.getTags());
+        if(image.getUser().getId()!=user.getId()){
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            model.addAttribute("deleteError",true);
+            return "images/image";}
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
